Refactor router
