+++ src/router/index.ts
@@ -2,9 +2,9 @@ import { createRouter, createWebHistory, createWebHashHistory } from 'vue-router
 // import HomeView from '@/views/HomeView.vue'
 // import ExamView from '@/views/ExamView.vue'
 // import ChapterView from '@/views/ChapterView.vue'
-import StudyView from '@/views/StudyView.vue'
-import LoginView from '@/views/LoginView.vue'
-import UserProfile from '@/views/UserProfile.vue'
+import AppView from '@/views/AppView.vue'
+import ExamView from '@/views/ExamView.vue'
+import WelcomeView from '@/views/WelcomeView.vue'
 import { EStudyView } from '@/types'
 
 const router = createRouter({
@@ -12,35 +12,21 @@ const router = createRouter({
   
   routes: [
     {
+      path: '/welcome',
+      name: 'login',
+      component: WelcomeView,
+    },
+    {
       path: '/',
       name: 'home',
-      component: StudyView,
-      meta: { type: EStudyView.Dashboard }
+      component: AppView,
+      meta: { type: EStudyView.Dashboard },
+      children: [
+        { path: 'exam/:exam',           component: ExamView, meta: { type: EStudyView.Exam } },
+        { path: 'exam/:exam/:chapter',  component: ExamView, meta: { type: EStudyView.Chapter } }
+      ]
     },
 
-    {
-      path: '/exam/:exam',
-      name: 'exam',
-      component: StudyView,
-      meta: { type: EStudyView.Exam }
-    },
-
-    {
-      path: '/exam/:exam/:chapter',
-      name: 'chapter',
-      component: StudyView,
-      meta: { type: EStudyView.Chapter }
-    },
-    {
-      path: '/login',
-      name: 'login',
-      component: LoginView,
-    },
-    {
-      path: '/profile',
-      name: 'profile',
-      component: UserProfile,
-    },
   ]
 })
