--- src/views/LoginView.vue
@@ -1,21 +1,0 @@
-<template>
-  <div>
-    <button @click="login">Login</button>
-    
-  </div>
-</template>
-<script lang="ts">
-  import { useAuth0 } from '@auth0/auth0-vue';
-
-  export default {
-    setup() {
-      const { loginWithRedirect } = useAuth0();
-
-      return {
-        login: () => {
-          loginWithRedirect();
-        }
-      };
-    }
-  };
-</script>
