+++ src/views/WelcomeView.vue
@@ -1,0 +1,9 @@
+<template>
+  <h1>Welcome to Study Buddy</h1>
+</template>
+
+<script setup lang="ts">
+</script>
+
+<style scoped lang="scss">
+</style>
